+++ src/main/java/com/example/review/domain/dto/HospitalResponse.java
@@ -10,17 +10,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class HospitalResponse {
-    private Integer id;
+    private Long id;
     private String roadNameAddress;
     private String hospitalName;
-    private Integer patientRoomCount;
-    private Integer totalNumberOfBeds;
+    private Long patientRoomCount;
+    private Long totalNumberOfBeds;
     private String businessTypeName;
     private Float totalAreaSize;
 
     private String businessStatusName;
 
-    public HospitalResponse(Integer id, String roadNameAddress, String hospitalName, Integer patientRoomCount, Integer totalNumberOfBeds, String businessTypeName, Float totalAreaSize) {
+    public HospitalResponse(Long id, String roadNameAddress, String hospitalName, Long patientRoomCount, Long totalNumberOfBeds, String businessTypeName, Float totalAreaSize) {
         this.id = id;
         this.roadNameAddress = roadNameAddress;
         this.hospitalName = hospitalName;
